fix: Rename copy phrase wrapper.
It used to be the signal_model now it is called copy phrase model
