+++ bcipy/tasks/rsvp/icon_to_icon.py
@@ -9,7 +9,7 @@ from psychopy import core
 from bcipy.display.rsvp.mode.icon_to_icon import IconToIconDisplay
 from bcipy.feedback.visual.visual_feedback import FeedbackType, VisualFeedback
 from bcipy.helpers.save import _save_session_related_data
-from bcipy.helpers.signal_model import CopyPhraseWrapper
+from bcipy.helpers.copy_phrase_wrapper import CopyPhraseWrapper
 from bcipy.helpers.task import (alphabet, generate_targets, get_user_input,
                                 process_data_for_decision,
                                 trial_complete_message)
